Split move PackageCRC-specific code from base.Package, remove unused
the code paths for ... were not reachable, removed

    SHIPPING_AUTH_CERTIFICATES
    SHIPPING_AUTH_IDENTITY
    SHIPPING_AUTH_S3_USERPASS
    SHIPPING_AUTH_USERPASS

methods only used by PackageCRC moved there
abstract methods unused outside PackageCRC removed

subclasses now need to provide .ship
